--- @navikt/core/react/src/layout/frame/index.ts
@@ -1,1 +1,0 @@
-export { Frame, type FrameProps } from "./Frame";
