+++ @navikt/core/react/src/layout/page/index.ts
@@ -1,0 +1,1 @@
+export { Page, type PageProps } from "./Page";
